requirements.txt
Flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
google-cloud-storage==2.16.0
google-cloud-secret-manager
google-genai
